\name{hivdata-package}
\alias{hivdata-package}
\docType{package}
\title{
 Six-Year Chronological Data of HIV and ART Cases in Pakistan
}

\description{
Infectious disease time series data are important to understand the pattern of diseases, timely intervention, and source allocation which helps in disease management and control. We provide the monthly number of HIV and ART cases of male, female, children and transgender as well as for the whole of Pakistan reported at various treatment centers in Pakistan from January 2016 to December 2021. The data have been taken from the official website of the Pakistan Bureau of Statistics. Every year the data is updated, Imran et al. (2018) used the monthly number of HIV cases from January 2013 to July 2014 taken from Pakistan Bureau of Statistics.
}

\details{
\tabular{ll}{
Package: \tab hivdata\cr
Type: \tab Package\cr
Version: \tab 0.1.0 \cr
Date: \tab 2023-10-07\cr
License: \tab GPL-2\cr
}
}


\author{
\itemize{
\item Muhammad Imran <imranshakoor84@yahoo.com>
\item Saima Shakoor <saimashakoor500@gmail.com>
}
}

\section{Maintainers}{
Muhammad Imran <imranshakoor84@yahoo.com>
}





\keyword{Time series data}
\keyword{HIV-AIDS}
\keyword{Antiretroviral therapy}





